fix: Missing AppSettings in NewTransferActivity when sharing files to the app before user login
If the user is logged out and shares files to the app, we won't find any AppSettings in Realm because they're initialized to an empty value during the login step.

A user will be logged out if the app received an intent before the user ever finished the onboarding. In case of such intent, we want to handle it correctly but we first need to connect the user seamlessly to avoid badly configured app settings