feat: Make upload copy-free
Now, the app starts a foreground service as soon as user-picked files
arrive in the app.
That way, we can keep the URI permissions until the upload
is complete, without having to import/copying the data
into the app storage. That lets us avoid filling the user device,
and avoids no free storage issues.

Change operations:

 1. Replace DB-backed upload session with in-memory UploadState
 2. Replace UploadWorker with UploadForegroundService
 3. Replace UploadFileTask with TransferUploader
 4. Replace TransferSendManager with UploadSessionStarter
 5. Replace ImportFilesScreen with PickFilesScreen
 6. Replace ImportFilesViewModel with PickFilesViewModel
 7. Merge all upload related screen (progress, errors, success, email validation…)
 8. Simplify navigation
 9. Remove ImportationFilesManager
10. Remove ImportLocalStorage
11. Ensure the old importation directory is removed if present, freeing storage
12. Extract a few classes into their own file
13. Update NotificationUtils code for naming semantics
14. Move files from the importfiles package to the pickfiles package